requirements.txt
pandas
openai
python-decouple
tenacity
